fix: prevent enter key from submitting forms
