+++ src/app/community/[communityId]/import-xlsx/page.tsx
@@ -5,6 +5,7 @@ import React from 'react';
 import { FormProvider } from '~/custom-hooks/hook-form';
 import { evictCache } from '~/graphql/apollo-client/cache-util/evict';
 import { appLabel, appPath } from '~/lib/app-path';
+import { Form } from '~/view/base/form';
 import { toast } from '~/view/base/toastify';
 import { ImportForm } from './import-form';
 import {
@@ -78,12 +79,12 @@ export default function ImportXlsx({ params }: RouteArgs) {
 
   return (
     <FormProvider {...formMethods}>
-      <form
+      <Form
         className="flex flex-col h-main-height"
         onSubmit={handleSubmit(onImport)}
       >
         <ImportForm />
-      </form>
+      </Form>
     </FormProvider>
   );
 }
