+++ src/app/community/[communityId]/property-list/community-modify-modal/modify-modal.tsx
@@ -7,6 +7,7 @@ import {
 } from '@nextui-org/react';
 import React from 'react';
 import { Button } from '~/view/base/button';
+import { Form } from '~/view/base/form';
 import { LastModified } from '~/view/last-modified';
 import { EventListEditor } from './event-list-editor';
 import { NameEditor } from './name-editor';
@@ -53,7 +54,7 @@ export const ModifyModal: React.FC<Props> = ({ hookForm, onSave }) => {
       isKeyboardDismissDisabled={true}
       hideCloseButton
     >
-      <form onSubmit={handleSubmit(onSubmit)}>
+      <Form onSubmit={handleSubmit(onSubmit)}>
         <ModalContent>
           <ModalHeader>Edit Community</ModalHeader>
           <ModalBody>
@@ -79,7 +80,7 @@ export const ModifyModal: React.FC<Props> = ({ hookForm, onSave }) => {
             </Button>
           </ModalFooter>
         </ModalContent>
-      </form>
+      </Form>
     </Modal>
   );
 };
