+++ src/app/community/[communityId]/property/[propertyId]/membership-editor-modal/modal-dialog.tsx
@@ -7,6 +7,7 @@ import {
 } from '@nextui-org/react';
 import React from 'react';
 import { Button } from '~/view/base/button';
+import { Form } from '~/view/base/form';
 import { MembershipInfoEditor } from './membership-info-editor';
 import { NotesEditor } from './notes-editor';
 import {
@@ -51,7 +52,7 @@ export const ModalDialog: React.FC<Props> = ({ hookForm, onSave }) => {
       isKeyboardDismissDisabled={true}
       hideCloseButton
     >
-      <form onSubmit={handleSubmit(onSubmit)}>
+      <Form onSubmit={handleSubmit(onSubmit)}>
         <ModalContent>
           <ModalHeader>Edit Membership Info</ModalHeader>
           <ModalBody>
@@ -71,7 +72,7 @@ export const ModalDialog: React.FC<Props> = ({ hookForm, onSave }) => {
             </Button>
           </ModalFooter>
         </ModalContent>
-      </form>
+      </Form>
     </Modal>
   );
 };
